+++ docs/reference/migration/migrate_3_0.asciidoc
@@ -203,7 +203,8 @@ cluster settings please use the settings update API and set their superseded key
 ==== Translog settings
 
 The `index.translog.flush_threshold_ops` setting is not supported anymore. In order to control flushes based on the transaction log
-growth use `index.translog.flush_threshold_size` instead.
+growth use `index.translog.flush_threshold_size` instead. Changing the translog type with `index.translog.fs.type` is not supported
+anymore, the `buffered` implementation is now the only available option and uses a fixed `8kb` buffer.
 
 [[breaking_30_mapping_changes]]
 === Mapping changes
